repair endpoint

+++ src/features/testing/infrastructure/testing.repository.ts
@@ -7,12 +7,19 @@ export class TestingRepository {
   constructor() {}
   async dropDb() {
     try {
-      const db = appSettings.api.MONGO_CONNECTION_URI;
-      //+
-      // '/' +
-      // appSettings.api.MONGO_DB_NAME;
-      await mongoose.connect(db); // Connecting to the database.
-      await mongoose.connection.db.dropDatabase();
+      let db: any;
+      if (appSettings.env.isProduction())
+        throw new Error('we need the rights of the creator');
+      if (appSettings.env.isTesting()) {
+        db = appSettings.api.MONGO_CONNECTION_URI_FOR_TESTS;
+      }
+      if (appSettings.env.isDevelopment()) {
+        db = appSettings.api.MONGO_CONNECTION_URI;
+      }
+      if (db) {
+        await mongoose.connect(db); // Connecting to the database.
+        await mongoose.connection.db.dropDatabase();
+      }
     } catch {
       console.log('DB dropping does failed');
       throw new Error('DB dropping did fail');
